requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.8.0
django-environ==0.6.0
django-heroku==0.3.1
django-spa==0.3.5
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
platformdirs==2.2.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
virtualenv==20.7.2
whitenoise==5.0.1
